pyproject.toml
[tool.poetry]
name = "proteus"
version = "0.1.0"
description = "A framework for analyzing protein structures using multiple deep learning models"
author = ["Jack Frazier <jnf44@cornell.edu> <jack.frazier03@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [
    {include = "src"},
    {include = "helpful", from = "."}
]
repository = "https://github.com/JackFrazier-jnf44/proteus"
keywords = [
    "protein",
    "structure",
    "prediction",
    "deep-learning",
    "bioinformatics",
    "structural-biology",
    "machine-learning",
    "openfold",
    "esm",
    "alphafold",
    "rosettafold"
]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.24.0"
torch = "^2.0.0"
biopython = "^1.81"
pandas = "^2.0.0"
matplotlib = "^3.7.0"
scikit-learn = "^1.3.0"
pydantic = "^2.0.0"
mmseqs2 = "^15.0.0"
openfold = "^1.0.0"
esm = "^2.0.0"
alphafold = "^2.3.0"
rosettafold = "^1.0.0"
colabfold = "^1.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
isort = "^5.12.0"
mypy = "^1.5.0"
ruff = "^0.1.0"
pytest-cov = "^4.1.0"

[tool.poetry.group.gpu]
optional = true

[tool.poetry.group.gpu.dependencies]
torch-cuda = "^2.0.0"
cupy = "^12.0.0"

[tool.poetry.scripts]
proteus = "src.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
multi_line_output = 3
known_first_party = ["src"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.ruff]
select = ["E", "F", "B", "I"]
ignore = ["E501"]
target-version = "py38" 